feat: Add conditional rendering for restaurant image in OrderStatusPage
This commit adds conditional rendering for the restaurant image in the OrderStatusPage component. If the order's restaurant has an image URL, it will be displayed using the AspectRatio component. Otherwise, a message indicating that no image is available will be shown. This enhancement improves the visual presentation of the order details and provides a better user experience.